renamed file, moved create_base_chart
